Package & main class rename
