+++ Export_Tool/src/main/java/com/gigaspaces/tools/importexport/FindSpaceClassTask.java
@@ -1,7 +1,7 @@
 /**
  * Aug 1, 2013
  */
-package com.gigaspaces.support.export;
+package com.gigaspaces.tools.importexport;
 
 import java.util.ArrayList;
 import java.util.List;
