+++ Export_Tool/src/main/java/com/gigaspaces/tools/importexport/SpaceClassExporter.java
@@ -1,7 +1,7 @@
 /**
  * Aug 1, 2013
  */
-package com.gigaspaces.support.export;
+package com.gigaspaces.tools.importexport;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
